Create test cases for VPDQ Index matching
