+++ python-threatexchange/threatexchange/extensions/vpdq/tests/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
